reducer 리팩토링
- store에 있던 reducer를 reducer 파일로 옮김
